requirements.txt
uvicorn
fastapi
datasets
transformers
torch
pillow
scikit-learn
torchvision